+++ Changelog.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `RecordWhenActive` to allow records to be deactivated during runtime, symbol `:WhenActive`
 * `RecordSubsolver` to record the result of a subsolver recording in the main solver, symbol `:Subsolver`
-* `RecordStoppingCriterion` to record the reason a solver stopped
+* `RecordStoppingReason` to record the reason a solver stopped
 * made the `RecordFactory` more flexible and quite similar to `DebugFactory`, such that it is now also easy to specify recordings at the end of solver runs. This can especially be used to record final states of sub solvers.
 
 ### Fixed